🚚 Move: TitleSection, ProfileImg 위치 이동
